Move the solver backends into src/solvers/
Fourteen files, seven backends. Public headers now install to
camada/solvers/, so consumers include <camada/solvers/z3solver.h>. That
is a breaking change, taken deliberately while camada is pre-1.0 rather
than keeping a flat include prefix that contradicts the tree.

PUBLIC_HEADER flattens every path into one destination, so the solver
headers are collected into their own list and installed separately.

Headers under solvers/ include their siblings as ../camadaexpr.h. Plain
names worked in-tree because the build adds src/ to the include path, but
they broke for consumers: the installed header sits in camada/solvers/,
where a quoted include resolves against that directory first and
camadaexpr.h is one level up. Caught by installing to a temp prefix and
compiling every installed header standalone from an empty include root --
worth doing directly, since the release-consumer CI job would otherwise
have been the first to notice.

The release consumer now includes a solver header too, so a future break
in this layout fails there rather than downstream. .gitignore also picks
up .cache/ and .vscode/, which were untracked editor state.
