+++ regression/bitwuzla.test.cpp
@@ -4,7 +4,7 @@
 #endif
 #include "tests.h"
 
-#include <bitwuzlasolver.h>
+#include <solvers/bitwuzlasolver.h>
 #include <catch2/catch_test_macros.hpp>
 #include <cstring>
 #include <random>
