+++ scripts/release-consumer/main.cpp
@@ -5,6 +5,12 @@
 #include <camada/camada.h>
 #include <camada/camadafeatures.h>
 
+// Exercise the solvers/ include path too, so a broken install layout for
+// the wrapper headers fails this consumer instead of a downstream build.
+#ifdef CAMADA_HAVE_Z3
+#include <camada/solvers/z3solver.h>
+#endif
+
 #include <cstdint>
 #include <iostream>
 #include <string>
