+++ src/solvers/bitwuzlasolver.cpp
@@ -22,9 +22,9 @@
 #if SOLVER_BITWUZLA_ENABLED
 
 #include "bitwuzlasolver.h"
-#include "camada.h"
-#include "camadacache.h"
-#include "camadacommon.h"
+#include "../camada.h"
+#include "../camadacache.h"
+#include "../camadacommon.h"
 
 #include <cstdio>
 #include <cstdlib>
