+++ src/solvers/yicessolver.cpp
@@ -22,8 +22,8 @@
 #if SOLVER_YICES_ENABLED
 
 #include "yicessolver.h"
-#include "camada.h"
-#include "camadacommon.h"
+#include "../camada.h"
+#include "../camadacommon.h"
 
 #include <cstdio>
 #include <gmp.h>
