+++ src/solvers/z3solver.cpp
@@ -22,8 +22,8 @@
 #if SOLVER_Z3_ENABLED
 
 #include "z3solver.h"
-#include "camada.h"
-#include "camadacommon.h"
+#include "../camada.h"
+#include "../camadacommon.h"
 
 #include <algorithm>
 #include <cassert>
